refactor(cache/redux-toolkit): redux-toolkit refactor done
