+++ packages/blockchain-wallet-v4-frontend/src/data/actions.ts
@@ -5,7 +5,7 @@ import { coreActions as core } from 'blockchain-wallet-v4/src'
 import { actions as alerts } from './alerts/slice'
 import * as analytics from './analytics/actions'
 import * as auth from './auth/actions'
-import * as cache from './cache/actions'
+import { actions as cache } from './cache/slice'
 import * as components from './components/actions'
 import * as custodial from './custodial/actions'
 import * as form from './form/actions'
