--- packages/blockchain-wallet-v4-frontend/src/data/cache/actionTypes.js
@@ -1,14 +1,0 @@
-export const ANNOUNCEMENT_DISMISSED = '@EVENT.ANNOUNCEMENT_DISMISSED'
-export const ANNOUNCEMENT_TOGGLED = '@EVENT.ANNOUNCEMENT_TOGGLED'
-export const CHANNEL_PRIV_KEY_CREATED = '@EVENT.CHANNEL_PRIV_KEY_CREATED'
-export const CHANNEL_CHANNEL_ID_CREATED = '@EVENT.CHANNEL_CHANNEL_ID_CREATED'
-export const CHANNEL_PHONE_CONNECTED = '@EVENT.CHANNEL_PHONE_CONNECTED'
-export const DISCONNECT_CHANNEL_PHONE = '@EVENT.DISCONNECT_CHANNEL_PHONE'
-// LAST ENTERED GUID, FROM OLDER VERSION OF LOGIN
-export const GUID_ENTERED = '@EVENT.GUID_ENTERED'
-
-// EMAIL LOGIN, INFO STORED FROM COOKIE
-export const EMAIL_STORED = '@EVENT.EMAIL_STORED'
-export const MOBILE_CONNECTED_STORED = '@EVENT.MOBILE_CONNECTED_STORED'
-export const GUID_STORED = '@EVENT.GUID_STORED'
-export const REMOVED_STORED_LOGIN = '@EVENT.REMOVE_STORED_LOGIN'
