+++ packages/blockchain-wallet-v4-frontend/src/data/rootReducer.ts
@@ -5,7 +5,7 @@ import { coreReducers, paths } from 'blockchain-wallet-v4/src'
 import { alertsReducer as alerts } from './alerts/slice'
 import { analyticsReducer } from './analytics/reducers'
 import auth from './auth/reducers'
-import cache from './cache/reducers'
+import { cacheReducer as cache } from './cache/slice'
 import components from './components/reducers'
 import { custodialReducer as custodial } from './custodial/reducers'
 import form from './form/reducers'
